+++ examples/game/test_performance.dart
@@ -33,7 +33,7 @@ main() async {
   String json = await _bundle.loadString('assets/sprites.json');
   _spriteSheet = new SpriteSheet(_images['assets/sprites.png'], json);
 
-  runApp(new App(
+  runApp(new MaterialApp(
     title: 'Test Sprite Performance',
     theme: _theme,
     routes: {
