+++ examples/game/test_physics.dart
@@ -27,7 +27,7 @@ main() async {
   String json = await _bundle.loadString('assets/sprites.json');
   _spriteSheet = new SpriteSheet(_images['assets/sprites.png'], json);
 
-  runApp(new App(
+  runApp(new MaterialApp(
     title: 'Test Physics',
     theme: new ThemeData(
       brightness: ThemeBrightness.light,
