+++ sky/packages/sky/lib/rendering.dart
@@ -23,6 +23,7 @@ export 'src/rendering/proxy_box.dart';
 export 'src/rendering/shifted_box.dart';
 export 'src/rendering/binding.dart';
 export 'src/rendering/stack.dart';
+export 'src/rendering/statistics_box.dart';
 export 'src/rendering/toggleable.dart';
 export 'src/rendering/view.dart';
 export 'src/rendering/viewport.dart';
